+++ ogcapi-types/src/styles/mod.rs
@@ -1,5 +1,5 @@
-mod mapbox;
-mod symcore;
+// mod mapbox;
+// mod symcore;
 
 use serde::{Deserialize, Serialize};
 use serde_json::Value;